Add base abs matcher for both implicit intent and explicit intent.
