+++ router/src/main/java/com/chenenyu/router/matcher/BrowserMatcher.java
@@ -1,8 +1,6 @@
 package com.chenenyu.router.matcher;
 
-import android.app.Activity;
 import android.content.Context;
-import android.content.Intent;
 import android.net.Uri;
 import android.support.annotation.Nullable;
 
@@ -14,7 +12,7 @@ import com.chenenyu.router.RouteOptions;
  * <p>
  * Created by Cheney on 2017/1/5.
  */
-public class BrowserMatcher extends AbsMatcher {
+public class BrowserMatcher extends AbsImplicitMatcher {
     public BrowserMatcher(int priority) {
         super(priority);
     }
@@ -23,10 +21,5 @@ public class BrowserMatcher extends AbsMatcher {
     public boolean match(Context context, Uri uri, @Nullable String route, RouteOptions routeOptions) {
         return !isEmpty(route) && (uri.toString().toLowerCase().startsWith("http://")
                 || uri.toString().toLowerCase().startsWith("https://"));
-    }
-
-    @Override
-    public Intent onMatched(Context context, Uri uri, @Nullable Class<? extends Activity> target) {
-        return new Intent(Intent.ACTION_VIEW, uri);
     }
 }
