+++ router/src/main/java/com/chenenyu/router/matcher/DirectMatcher.java
@@ -1,19 +1,17 @@
 package com.chenenyu.router.matcher;
 
-import android.app.Activity;
 import android.content.Context;
-import android.content.Intent;
 import android.net.Uri;
 import android.support.annotation.Nullable;
 
 import com.chenenyu.router.RouteOptions;
 
 /**
- * Internal default matcher.
+ * Absolutely matcher.
  * <p>
  * Created by Cheney on 2016/12/23.
  */
-public class DirectMatcher extends AbsMatcher {
+public class DirectMatcher extends AbsExplicitMatcher {
 
     public DirectMatcher(int priority) {
         super(priority);
@@ -22,14 +20,6 @@ public class DirectMatcher extends AbsMatcher {
     @Override
     public boolean match(Context context, Uri uri, @Nullable String route, RouteOptions routeOptions) {
         return !isEmpty(route) && uri.toString().equals(route);
-    }
-
-    @Override
-    public Intent onMatched(Context context, Uri uri, @Nullable Class<? extends Activity> target) {
-        if (target == null) {
-            return null;
-        }
-        return new Intent(context, target);
     }
 
 }
